+++ rdagent/scenarios/data_science/proposal/exp_gen/ckp_select.py
@@ -65,7 +65,6 @@ class LimitTimeCKPSelector(CheckpointSelector):
         current_time = datetime.now()
 
         if len(trace.hist) == 0:
-            trace.sub_trace_count = 0
             self.sub_trace_start_times[trace.sub_trace_count] = current_time
             logger.info(f"Starting initial sub-trace {trace.sub_trace_count} at {current_time}")
             return (-1,)  # Continue with latest trial for new sub-trace
@@ -90,12 +89,11 @@ class LimitTimeCKPSelector(CheckpointSelector):
                 return (-1,)
 
             # Time limit exceeded, start a new sub-trace
-            trace.sub_trace_count += 1
-            self.sub_trace_start_times[trace.sub_trace_count] = current_time
+            self.sub_trace_start_times[trace.sub_trace_count + 1] = current_time
             logger.info(
                 f"Elapsed time {elapsed_time} exceeds time limit {self.time_limit_pre_trace}, jump to a new sub-trace"
             )
-            logger.info(f"current sub-trace count: {trace.sub_trace_count}")
+            logger.info(f"current sub-trace count: {trace.sub_trace_count + 1}")
             return tuple()  # Empty tuple signals starting a new sub-trace
 
 
@@ -138,11 +136,10 @@ class SOTAJumpCKPSelector(CheckpointSelector):
                     logger.info(f"current sub-trace count: {trace.sub_trace_count}")
                     return (-1,)
 
-                trace.sub_trace_count += 1
                 logger.info(
                     f"SOTA count {sota_count} is below threshold {self.SOTA_COUNT_THRESHOLD}, jump to a new sub-trace"
                 )
-                logger.info(f"current sub-trace count: {trace.sub_trace_count}")
+                logger.info(f"current sub-trace count: {trace.sub_trace_count + 1}")
                 return ()
             else:
                 logger.info(
@@ -201,7 +198,6 @@ class BackJumpCKPSelector(CheckpointSelector):
 
                 random_choice = random.random()
                 if random_choice < 0.5:
-                    trace.sub_trace_count += 1
                     logger.info(
                         f"SOTA count {sota_count} is below threshold {self.SOTA_COUNT_THRESHOLD}, jump a new sub-trace"
                     )
@@ -227,11 +223,10 @@ class BackJumpCKPSelector(CheckpointSelector):
                             logger.info(f"current sub-trace count: {trace.sub_trace_count}")
                             return (-1,)
 
-                        trace.sub_trace_count += 1
                         logger.info(
                             f"SOTA count {sota_count} is below threshold {self.SOTA_COUNT_THRESHOLD}, jump a new sub-trace"
                         )
-                        logger.info(f"current sub-trace count: {trace.sub_trace_count}")
+                        logger.info(f"current sub-trace count: {trace.sub_trace_count + 1}")
                         return ()  # reboot a new sub-trace
 
             else:
