+++ rdagent/scenarios/data_science/proposal/exp_gen/merge.py
@@ -8,13 +8,13 @@ from rdagent.core.proposal import ExpGen
 from rdagent.log import rdagent_logger as logger
 from rdagent.log.timer import RD_Agent_TIMER_wrapper, RDAgentTimer
 from rdagent.scenarios.data_science.experiment.experiment import DSExperiment
-from rdagent.scenarios.data_science.proposal.exp_gen import DSExpGen
+from rdagent.scenarios.data_science.loop import DataScienceRDLoop
 from rdagent.scenarios.data_science.proposal.exp_gen.base import DSHypothesis, DSTrace
 from rdagent.utils.agent.tpl import T
 
 
 class MergeExpGen(ExpGen):
-    def gen(self, trace: DSTrace, selection: tuple[int, ...] = (-1,)) -> DSExperiment:
+    def gen(self, trace: DSTrace) -> DSExperiment:
         # Ignore the selection argument and use all leaves instead.
         leaves: list[int] = trace.get_leaves()
         trace.set_current_selection((leaves[0],))  # override the current selection.
@@ -87,9 +87,11 @@ class ExpGen2TraceAndMerge(ExpGen):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.merge_exp_gen = MergeExpGen(self.scen)
-        self.exp_gen = DSExpGen(self.scen)
+        self.exp_gen = DataScienceRDLoop._get_exp_gen(
+            "rdagent.scenarios.data_science.proposal.exp_gen.DSExpGen", self.scen
+        )
 
-    def gen(self, trace: DSTrace, selection: tuple[int, ...] = (-1,)) -> DSExperiment:
+    def gen(self, trace: DSTrace) -> DSExperiment:
         timer: RDAgentTimer = RD_Agent_TIMER_wrapper.timer
         logger.info(f"Remain time: {timer.remain_time_duration}")
 
@@ -101,24 +103,23 @@ class ExpGen2TraceAndMerge(ExpGen):
                 selection = (
                     leaves[0],
                 )  # continue the first trace. This will result in the interleaving of two traces expansion.
-            return self.exp_gen.gen(trace, selection)
+            trace.set_current_selection(selection)
+            return self.exp_gen.gen(trace)
         else:
             # disable reset in merging stage
             DS_RD_SETTING.coding_fail_reanalyze_threshold = 100000
             DS_RD_SETTING.consecutive_errors = 100000
 
-            leaves: list[int] = trace.get_leaves()
-            if len(leaves) < 2:
-                return self.exp_gen.gen(trace, selection)
+            if trace.sub_trace_count < 2:
+                return self.exp_gen.gen(trace)
             else:
-                return self.merge_exp_gen.gen(trace, selection)
+                return self.merge_exp_gen.gen(trace)
 
 
 class MergeExpGen_MultiTrace(ExpGen):
-    def gen(self, trace: DSTrace, selection: tuple[int, ...] = (-1,)) -> DSExperiment:
+    def gen(self, trace: DSTrace) -> DSExperiment:
         # Ignore the selection argument and use all leaves instead.
         leaves: list[int] = trace.get_leaves()
-        trace.set_current_selection(selection)  #
 
         # assuming merging the first and sencond trace.
         sota_exp_fb = trace.sota_experiment_fb(selection=(leaves[0],))
@@ -195,11 +196,13 @@ class ExpGen2TraceAndMergeV2(ExpGen):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.merge_exp_gen = MergeExpGen_MultiTrace(self.scen)
-        self.exp_gen = DSExpGen(self.scen)
+        self.exp_gen = DataScienceRDLoop._get_exp_gen(
+            "rdagent.scenarios.data_science.proposal.exp_gen.DSExpGen", self.scen
+        )
         self.MAX_TRACE_NUM = DS_RD_SETTING.max_trace_num  # maximum number of traces to grow before merging
         self.flag_start_merge = False
 
-    def gen(self, trace: DSTrace, selection: tuple[int, ...] = (-1,)) -> DSExperiment:
+    def gen(self, trace: DSTrace) -> DSExperiment:
         timer: RDAgentTimer = RD_Agent_TIMER_wrapper.timer
         logger.info(f"Remain time: {timer.remain_time_duration}")
 
@@ -214,8 +217,7 @@ class ExpGen2TraceAndMergeV2(ExpGen):
                 else:
                     # set the knowledge base option back to False for the other traces
                     DS_RD_SETTING.enable_knowledge_base = False
-
-            return self.exp_gen.gen(trace, selection)
+            return self.exp_gen.gen(trace)
 
         else:
             # disable reset in merging stage
@@ -224,15 +226,14 @@ class ExpGen2TraceAndMergeV2(ExpGen):
 
             leaves: list[int] = trace.get_leaves()
             if len(leaves) < 2:
-                return self.exp_gen.gen(trace, selection=(-1,))
+                trace.set_current_selection(selection=(-1,))
+                return self.exp_gen.gen(trace)
             else:
-
                 if not self.flag_start_merge:  # root node of the merge trace
                     self.flag_start_merge = True
-                    selection = tuple()
-                    return self.merge_exp_gen.gen(trace, selection)
+                    trace.set_current_selection(tuple())
+                    return self.merge_exp_gen.gen(trace)
                 else:
-                    # return self.merge_exp_gen.gen(trace, selection)
-                    return self.exp_gen.gen(
-                        trace, selection=(-1,)
-                    )  # continue the last trace, to polish the merged solution
+                    # return self.merge_exp_gen.gen(trace)
+                    trace.set_current_selection(selection=(-1,))
+                    return self.exp_gen.gen(trace)  # continue the last trace, to polish the merged solution
